CORE-9210 Move Subject sorting to callback
